+++ Source/Clients/DotNET/Events/EventSourceIdJsonConverterFactory.cs
@@ -6,8 +6,6 @@ using System.Text.Json.Serialization;
 
 namespace Cratis.Chronicle.Events;
 
-#pragma warning disable SA1402,MA0048 // File may only contain a single type
-
 /// <summary>
 /// Represents a <see cref="JsonConverterFactory"/> that produces <see cref="EventSourceIdJsonConverter{T}"/>
 /// instances for any closed <see cref="EventSourceId{T}"/> type.
@@ -25,28 +23,5 @@ public class EventSourceIdJsonConverterFactory : JsonConverterFactory
         var typeArg = typeToConvert.GetGenericArguments()[0];
         var converterType = typeof(EventSourceIdJsonConverter<>).MakeGenericType(typeArg);
         return (Activator.CreateInstance(converterType) as JsonConverter)!;
-    }
-}
-
-/// <summary>
-/// Represents a <see cref="JsonConverter{T}"/> for <see cref="EventSourceId{T}"/> that serializes
-/// the typed value as a plain string for wire compatibility with <see cref="EventSourceId"/>.
-/// </summary>
-/// <typeparam name="T">The underlying type wrapped by <see cref="EventSourceId{T}"/>.</typeparam>
-public class EventSourceIdJsonConverter<T> : JsonConverter<EventSourceId<T>>
-    where T : IComparable
-{
-    /// <inheritdoc/>
-    public override EventSourceId<T>? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-    {
-        var value = reader.GetString();
-        return value is null ? null : (EventSourceId<T>)value;
-    }
-
-    /// <inheritdoc/>
-    public override void Write(Utf8JsonWriter writer, EventSourceId<T> value, JsonSerializerOptions options)
-    {
-        EventSourceId id = value;
-        writer.WriteStringValue(id.Value);
     }
 }
